Remove dead code and improve code quality
- Remove unused WiFiManager::handle() and _instance
- Remove unused TemperatureSensor::isConnected() and getLastReadTime()
- Remove unused MQTTManager::getClient()
- Remove unused OTAManager::setPassword() and setHostname()
- Fix duplicate header guard in TemperatureSensor.h
- Change InputManager pin types from int to uint8_t for consistency
- Add comprehensive CODE_QUALITY_REPORT.md

Estimated flash savings: ~500 bytes
RAM savings: 4-8 bytes

+++ src/InputManager.cpp
@@ -7,7 +7,7 @@ InputManager::InputManager()
 {
 }
 
-void InputManager::begin(int buttonPin, int ledPin, unsigned long debounceDelay, unsigned long ledBlinkInterval)
+void InputManager::begin(uint8_t buttonPin, uint8_t ledPin, unsigned long debounceDelay, unsigned long ledBlinkInterval)
 {
   _buttonPin = buttonPin;
   _ledPin = ledPin;
